app/backend/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
drf-registration
django-cors-headers==3.5.0
phonenumbers==8.12.15
django-nested-admin
psycopg2